Revert util file changes
